refactor: Divide code into multiple files
