+++ src/components/visualization/ChartTableWrapper.tsx
@@ -1,6 +1,6 @@
 import { DataRecord, Resource } from '@/types/visualization';
 import BarChart from './BarChart';
-import ChartTableFilter from './ChartTableFilter';
+import ChartTableFilter from './chart-table-filter/ChartTableFilter';
 import Table from './Table';
 import { useState } from 'react';
 import useWindowDimensions from '../helper/WindowDimensions';
