+++ src/components/visualization/chart-table-filter/InputWithFloatingLabel.tsx
@@ -1,0 +1,22 @@
+import { ChangeEvent } from 'react';
+
+export function InputWithFloatingLabel({
+  id,
+  type,
+  value,
+  label,
+  onChange,
+}: {
+  id: string;
+  type: string;
+  value: string;
+  label: string;
+  onChange: (e: ChangeEvent<HTMLInputElement>) => void;
+}) {
+  return (
+    <div className="form-floating">
+      <input type={type} id={id} className="form-control rounded-0" value={value} onChange={onChange} />
+      <label htmlFor={id}>{label}</label>
+    </div>
+  );
+}
